refactor: rename tile module to utilities
