+++ src/geometries/geometry.ts
@@ -2,8 +2,8 @@ import type { BBox, Position } from 'geojson';
 import { DEFAULT_CRS } from '../constants';
 import { decodeFromJSON, encodeToJSON } from '../crs/crs';
 import type { TileMatrixSet } from '../tiles/tileMatrixSet';
-import { positionToTileIndex, tileMatrixToBBox } from '../tiles/tiles';
 import type { CRS as CRSType, TileIndex } from '../tiles/types';
+import { positionToTileIndex, tileMatrixToBBox } from '../tiles/utilities';
 import type { ArrayElement, CoordRefSysJSON } from '../types';
 import { validateCRS, validateCRSByOtherCRS, validateMetatile } from '../validations/validations';
 import type { GeoJSONBaseGeometry, GeoJSONGeometry, JSONFGFeature } from './types';
