+++ tests/geometries/point.spec.ts
@@ -5,8 +5,8 @@ import { Point } from '../../src/geometries/point';
 import type { GeoJSONPoint, PointInput } from '../../src/geometries/types';
 import { TileMatrixSet } from '../../src/tiles/tileMatrixSet';
 import { TILEMATRIXSETJSON_WORLD_CRS84_QUAD } from '../../src/tiles/tileMatrixSets/worldCRS84Quad';
-import { tileMatrixToBBox } from '../../src/tiles/tiles';
 import type { TileMatrixLimits, TileMatrixSetJSON, TileMatrixSet as TileMatrixSetType } from '../../src/tiles/types';
+import { tileMatrixToBBox } from '../../src/tiles/utilities';
 import { generatePointInput } from './helpers/geometries';
 import { generateNonFinite, isSafeInteger } from './helpers/propertyTest';
 import { generateTileMatrixToBBox } from './helpers/tiles';
